+++ incubacloud/models/__init__.py
@@ -1,3 +1,4 @@
+from . import http_session_cookie
 from . import cloud_security_mixin
 from . import cloud_audit_tracked_mixin
 from . import cloud_pip_provenance_mixin